fix: Inital pass on final code review.
