+++ src/lib/interaction/layer.ts
@@ -1,4 +1,7 @@
 import type { Map } from 'maplibre-gl';
+import type { FeatureCollection } from 'geojson';
+import uniqueId from 'lodash.uniqueid';
+import kebabCase from 'lodash.kebabcase';
 
 import { deriveColorScale } from '$lib/interaction/color';
 import { deriveSize } from '$lib/interaction/geometry';
@@ -11,6 +14,14 @@ import {
   instrumentPointSelect
 } from '$lib/interaction/select';
 import type { CartoKitLayer } from '$lib/types/CartoKitLayer';
+import { randomColor } from '$lib/utils/color';
+import {
+  DEFAULT_OPACITY,
+  DEFAULT_RADIUS,
+  DEFAULT_STROKE_OPACITY,
+  DEFAULT_STROKE_WIDTH
+} from '$lib/utils/constants';
+import { getLayerGeometryType } from '$lib/utils/geojson';
 
 /**
  * Add a CartoKit layer to the map.
@@ -194,3 +205,75 @@ export function addLayer(map: Map, layer: CartoKitLayer): void {
     }
   }
 }
+
+/**
+ * Generate a CartoKitLayer for a given GeoJSON dataset, using the dataset's
+ * Geometry type to select the appropriate member from the CartoKitLayer union.
+ *
+ * @param featureCollection – The GeoJSON FeatureCollection associated with the
+ * layer.
+ * @returns – A default CartoKitLayer appropriate for the input geometry type.
+ */
+export const generateCartoKitLayer = (
+  featureCollection: FeatureCollection,
+  options: { displayName: string; url?: string; fileName?: string }
+): CartoKitLayer => {
+  const geometryType = getLayerGeometryType(featureCollection);
+  const color = randomColor();
+
+  switch (geometryType) {
+    case 'Point':
+    case 'MultiPoint':
+      return {
+        id: uniqueId(`${kebabCase(options.displayName)}__`),
+        displayName: options.displayName,
+        type: 'Point',
+        data: {
+          geoJSON: featureCollection,
+          rawGeoJSON: featureCollection,
+          fileName: options.fileName,
+          url: options.url,
+          transformations: []
+        },
+        style: {
+          size: DEFAULT_RADIUS,
+          fill: {
+            color,
+            opacity: DEFAULT_OPACITY
+          },
+          stroke: {
+            color,
+            width: DEFAULT_STROKE_WIDTH,
+            opacity: DEFAULT_STROKE_OPACITY
+          }
+        }
+      };
+    case 'Polygon':
+    case 'MultiPolygon':
+      return {
+        id: uniqueId(`${kebabCase(options.displayName)}__`),
+        displayName: options.displayName,
+        type: 'Fill',
+        data: {
+          geoJSON: featureCollection,
+          rawGeoJSON: featureCollection,
+          fileName: options.fileName,
+          url: options.url,
+          transformations: []
+        },
+        style: {
+          fill: {
+            color,
+            opacity: DEFAULT_OPACITY
+          },
+          stroke: {
+            color,
+            width: DEFAULT_STROKE_WIDTH,
+            opacity: DEFAULT_STROKE_OPACITY
+          }
+        }
+      };
+    default:
+      throw new Error(`Unsupported geometry type: ${geometryType}.`);
+  }
+};
